feat: adapted transition of the Photos, Files and Home page. Need final robustness corrections.
Robustness corrections needed :
Protection of click while transitionning
Hover effects are often reset during the transition, due to the CustomRouter changing the rendered components.
Causes the clicked widget on the home page to change size abruptly.
It also happens when arriving on the Home page, if the mouse hovers a widget.